requirements-tests.txt
#Faker==8.10.1
Pillow==9.0.1
django-analytical==3.1.0
djangorestframework==3.12.4